src/requirements.txt
PyICU
numpy
six
pycld2
morfessor
polyglot
